+++ tests/CodeigniterModelTest.php
@@ -125,6 +125,19 @@ class CodeigniterModelTest extends \PHPUnit_Framework_TestCase
     }
 
     /**
+     * Tests CodeigniterModel::validation method.
+     *
+     * @return void
+     */
+    public function testValidateMethod()
+    {
+        $expected  = [ 'name' => 'The Name field is required.' ];
+        $validated = $this->ci->comment->validate([ 'message' => 'test' ]);
+
+        $this->assertEquals($expected, $this->ci->comment->validation_errors());
+    }
+
+    /**
      * Tests CodeigniterModel::paginate method.
      *
      * @return void